MiKo_2042 no longer reports on '<br/>'
see https://learn.microsoft.com/en-us/dotnet/csharp/language-reference/xmldoc/recommended-tags#br

(fixes #152)
