+++ MiKo.Analyzer.Shared/Resources.resx
@@ -2353,19 +2353,19 @@ This format clearly describes the purpose of the class and its functionality, ma
     <value>&lt;summary&gt; documentation should not contain other documentation tags</value>
   </data>
   <data name="MiKo_2042_CodeFixTitle" xml:space="preserve">
-    <value>Replace &lt;br/&gt; with &lt;para/&gt;</value>
+    <value>Replace &lt;p&gt; with &lt;para&gt;</value>
   </data>
   <data name="MiKo_2042_Description" xml:space="preserve">
-    <value>Documentation should use the '&lt;para/&gt;' XML tags instead of '&lt;br/&gt;' or '&lt;p/&gt;' HTML tags. This ensures a consistent format, suitable for XML-based documentation tools and frameworks.</value>
+    <value>Documentation should use the '&lt;para&gt;' XML tags instead of '&lt;p&gt;' HTML tags. This ensures a consistent format, suitable for XML-based documentation tools and frameworks.</value>
   </data>
   <data name="MiKo_2042_HelpLinkUri" xml:space="preserve">
     <value>https://learn.microsoft.com/en-us/dotnet/csharp/programming-guide/xmldoc/para</value>
   </data>
   <data name="MiKo_2042_MessageFormat" xml:space="preserve">
-    <value>Use '&lt;para/&gt;' instead of '{0}'</value>
+    <value>Use '&lt;para&gt;' instead of '&lt;p&gt;'</value>
   </data>
   <data name="MiKo_2042_Title" xml:space="preserve">
-    <value>Documentation should use '&lt;para/&gt;' XML tags instead of '&lt;br/&gt;' HTML tags</value>
+    <value>Documentation should use '&lt;para&gt;' XML tags instead of '&lt;p&gt;' HTML tags</value>
   </data>
   <data name="MiKo_2043_CodeFixTitle" xml:space="preserve">
     <value>Start &lt;summary&gt; with 'Encapsulates a method that '</value>
